Implemented test for BZ1428690
